pyproject.toml
[project]
name = "star-find-your-spark-build-your-future"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.32.0",
    "pdfplumber>=0.11.7",
    "psycopg2-binary>=2.9.10",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.43",
    "streamlit>=1.49.0",
]
